+++ src/common/utils/hash-code.ts
